added fortmatic, fixed ledger timeout
